Create a building placement algorithm
I was going for a spiral algorithm like you see with the Minecraft "loading
world screen" but there was a bug in my code and it generated a shape of a swan.
It makes me happy so I kept it. It's not a bug - it's an epic feature.

PHP runtimes are now officially modeled as "Swan Town".